feat(urm/registry): mom, raven, hopfield, conformer → native tier (stale label fix)
Four reference-tier rows already ran native correctly — their tier labels were stale
from before the underlying ops went native:

- mom (delegates to GatedDeltaNetLayer → K2 matrix_scan, native since the A8 admissions)
- raven (_SlotAttentionBase → internal K2LinearStateLayer, native since the K2 native tier)
- hopfield_association (HopfieldAssociationLayer → K1 online softmax; the fp32 `scaling`
  parameter promoted the query to fp32 under autocast — cast back to the projection dtype)
- conformer_attention (ConformerRelPosAttention → K1 online softmax with score_bias; the
  fp32 pos_bias_u promoted the query, and score_bias must be fp32 explicitly)

Registry: 42 of 51 rows native. The 9 remaining reference rows are all kernel work:
K1 law branches (wall/kata/pattention/tda), wide-head K1 (mla/tucker), native merge op
(differential), K2 elementwise gate (mamba1), K2 normalized (based_attention).

Suite: 747 passed, 40 skipped, 0 failed.

+++ projects/urm/architectures/conformer_attention.py
@@ -141,11 +141,13 @@ class ConformerRelPosAttention(torch.nn.Module):
 
         # Mixer: q carries pos_bias_u; the K1 key-dim scale completes 1/√d.
         # The pinned mask is "True = masked out"; the K1 attention_mask role is
-        # "True = visible", so invert it here.
-        q_with_bias_u = q + self.pos_bias_u
+        # "True = visible", so invert it here. The native K1 kernel requires one
+        # dtype across q/k/v and score_bias in fp32 — cast accordingly.
+        q_with_bias_u = (q + self.pos_bias_u).to(q.dtype)
         visible = ~mask if mask is not None else None
         out = self._plan.execute(
-            query=q_with_bias_u, key=k, value=v, score_bias=score_bias,
+            query=q_with_bias_u, key=k, value=v,
+            score_bias=score_bias.float(),
             attention_mask=visible,
         )["output"]
         return self.linear_out(out.reshape(B, T1, self.num_heads * self.d_k))
